PortfolioStockService/requirements.txt
Flask==3.0.0
flask-restful==0.3.10
pymongo==4.10.1
python-dotenv==1.0.1